+++ src/main/java/com/xx/pojo/Blog.java
@@ -3,6 +3,9 @@ package com.xx.pojo;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableLogic;
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import com.xx.util.FormatTime;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,20 +15,21 @@ import java.sql.Timestamp;
 import java.text.ParseException;
 import java.util.List;
 import java.util.Map;
+
 @Data
 @NoArgsConstructor
 @TableName
 
 public class Blog {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String cover;
     private String content;
-    private Integer up;
-    private Integer down;
-    private Integer star;
-    private Integer views;
+    private Long up;
+    private Long down;
+    private Long star;
+    private Long views;
     private String authorUsername;
     private Long postTime;
     private Boolean status;
